dbt test results not updated [sc-30043]
